$ Some more adjustments to the package structure
