+++ src/com/blazeloader/api/client/gui/IModInventory.java
@@ -1,4 +1,4 @@
-package com.blazeloader.api.client.api.gui;
+package com.blazeloader.api.client.gui;
 
 import net.minecraft.inventory.IInventory;
 import net.minecraft.world.IInteractionObject;
